+++ src/inspector/qml/MuseScore/Inspector/measures/measuressettingsmodel.h
@@ -40,7 +40,7 @@ class MeasuresSettingsModel : public AbstractInspectorModel
     Q_PROPERTY(bool allSystemsAreLocked READ allSystemsAreLocked NOTIFY allSystemsAreLockedChanged)
     Q_PROPERTY(bool scoreIsInPageView READ scoreIsInPageView NOTIFY scoreIsInPageViewChanged)
     Q_PROPERTY(bool isMakeIntoSystemAvailable READ isMakeIntoSystemAvailable NOTIFY isMakeIntoSystemAvailableChanged)
-    Q_PROPERTY(size_t systemCount READ systemCount NOTIFY systemCountChanged)
+    Q_PROPERTY(int systemCount READ systemCount NOTIFY systemCountChanged)
 
 public:
     explicit MeasuresSettingsModel(QObject* parent, IElementRepositoryService* repository);
@@ -80,7 +80,7 @@ public:
     bool scoreIsInPageView() const;
     bool isMakeIntoSystemAvailable() const;
 
-    size_t systemCount() const;
+    int systemCount() const;
 
 protected:
     void onNotationChanged(const mu::engraving::PropertyIdSet&, const mu::engraving::StyleIdSet&) override;
@@ -95,7 +95,7 @@ signals:
     void allSystemsAreLockedChanged(bool allLocked);
     void scoreIsInPageViewChanged(bool isInPageView);
     void isMakeIntoSystemAvailableChanged(bool isMakeIntoSystemAvailable);
-    void systemCountChanged(size_t count);
+    void systemCountChanged(int count);
 
 private:
     bool m_allSystemsAreLocked = false;
